Added Rest Controller interfaces
